+++ src/app.mts
@@ -1,0 +1,7 @@
+import { cubicFireConfigs } from "./apps/cubic-fire.mjs";
+import { fractalConfigs } from "./apps/fractal.mjs";
+
+// const appConfigs = cubicFireConfigs;
+const appConfigs = fractalConfigs;
+
+export default appConfigs;
