+++ src/main.mts
@@ -8,23 +8,23 @@ import {
   renderLagopusTree,
   resetCanvasHeight,
   computeBasePoints,
-  createGlobalPointsBuffer,
-  BaseCellParams,
 } from "./index.mjs";
 
-import movePoints from "../shaders/move-points.wgsl";
-import fireCompute from "../shaders/fire-compute.wgsl";
-import strokeWgsl from "../shaders/stroke.wgsl";
-import cubicFire from "../shaders/cubic-fire.wgsl";
-import { useBaseSize, useRemoteControl } from "./config.mjs";
-import { Number4, rand, randBalance } from "./math.mjs";
+import { useRemoteControl } from "./config.mjs";
+import { Atom } from "./atom.mjs";
+
+import appConfigs from "./app.mjs";
 
 let canvas = document.querySelector("canvas");
 let timeoutState: NodeJS.Timeout;
 let rafState = 0;
 
+let computeShaderState = new Atom(appConfigs.computeShader);
+
 let loopPaint = () => {
-  computeBasePoints(fireCompute);
+  if (computeShaderState.value) {
+    computeBasePoints(computeShaderState.value);
+  }
   paintLagopusTree();
   timeoutState = setTimeout(() => {
     rafState = requestAnimationFrame(loopPaint);
@@ -32,24 +32,10 @@ let loopPaint = () => {
   // rafState = requestAnimationFrame(loopPaint);
 };
 
-let createBasePoint = (idx: number): BaseCellParams => {
-  let offset = 200;
-  let armOffset = 120;
-  let position: Number4 = [randBalance(offset), randBalance(offset), randBalance(offset), 1];
-  let velocity: Number4 = [0, 3 + randBalance(3), 0, 0];
-  // let arm: Number4 = [randBalance(armOffset), randBalance(armOffset), randBalance(armOffset), 1];
-  // let arm: Number4 = [100, 0, 0, 0];
-  let arm: Number4 = [randBalance(armOffset), 0, randBalance(armOffset), 0];
-  let params: Number4 = [rand(10), 2 + rand(2), 0, 0];
-  let extendParams: Number4 = [idx, idx, idx, idx];
-  return { position, arm, velocity, params, extendParams };
-};
-
 window.onload = async () => {
   await initializeContext();
-  createGlobalPointsBuffer(useBaseSize, createBasePoint);
-  // renderLagopusTree(strokeWgsl);
-  renderLagopusTree(cubicFire);
+  appConfigs.initPointsBuffer();
+  renderLagopusTree(appConfigs.renderShader);
   loadTouchControl();
 
   loopPaint();
@@ -65,13 +51,12 @@ window.onload = async () => {
   }
 };
 
-import.meta.hot?.accept("./app/container", (container) => {
-  createGlobalPointsBuffer(useBaseSize, createBasePoint);
-  clearTimeout(timeoutState);
-  cancelAnimationFrame(rafState);
+import.meta.hot?.accept("./app.mts", (app) => {
+  let newAppConfigs = app.default as typeof appConfigs;
+  newAppConfigs.initPointsBuffer();
 
-  renderLagopusTree(container.compContainer());
-  loopPaint();
+  renderLagopusTree(newAppConfigs.renderShader);
+  computeShaderState.value = newAppConfigs.computeShader;
 
   console.log("reloaded");
 });
